New implementation of conduit-lwt-unix with Lwt's channel, tls and ssl layers
This is the first real implementation of `conduit` with some protocols:
- tcp/ip with `Lwt_unix`
- tls with `conduit-tls`
- ssl with `Lwt_ssl`

Due to the ability to use `Lwt_io`, this package provides an helper to
give `Lwt_io.channel` from an abstracted `Conduit_lwt{_unix}.flow`.

It provides TLS + provided tcp/ip protocol, a possible composition
with SSL layer (with `Lwt_ssl`) and SSL + provided tcp/ip protocol.

It provides a `resolv_conf` function as the main DNS resolver. It's a
call of `gethostbyname` and finally trusts on your `resolv.conf`.