New implementation of conduit-tls which provides a composition mechanism of the TLS layer with another protocol
`ocaml-tls` is a new package which provides internally a _functor_
to compose a Conduit.Sigs.PROTOCOL with TLS. It gives an example
of composition of protocols in `conduit`.

It requires an implementation of `conduit` which, at least, uses
`Cstruct.t` as `input` and `output`.

A realistic example is the composition of tcp/ip protocols provided by:
- `conduit-lwt-unix`
- `conduit-async`
- `conduit-mirage`

Each of them are differents (about type and implementation) but they
use this package to provide a TLS layer on top of them. At the end,
composition should be less than 10 lines of code.

Due to the non ability to use the scheduler, provided implementation
is not protected against data-race condition. The documentation tells
you more about that.